terminal implementation/requirements.txt
phe==1.4.0
pycryptodome==3.18.0
pickle5==0.0.13
